requirements.txt
edge_tts==6.1.12
httpx==0.27.0
numpy==2.0.0
protobuf==5.27.2
PyAudioWPatch==0.2.12.6
PyQt5==5.15.10
PyQt5_sip==12.13.0
Requests==2.32.3
scipy==1.14.0
SoundCard==0.4.3
torch==2.3.1
torchaudio==2.3.1
webrtcvad==2.0.10
google-cloud-speech
